Move App to own folder within components
